web_app/requirements.txt
Flask==1.1.2
pandas
gcsfs
google-cloud-storage
google-api-python-client
google-api-core
google-auth==1.19.2
requests==2.24.0
